[UPDATE] Renamed package to group package together and make it clear
Because you can't add medium
